refactor: 清理未使用的导入和优化代码结构
- 移除多个模块中未使用的导入语句
- 优化WhitelistConfig的Default实现
- 简化NfqConfig字段命名并标记为未使用
- 清理pipeline模块中的未使用代码
- 优化main.rs中未使用的函数参数
- 改进crypto模块中的哈希计算参数传递
- 移除state模块中未使用的导入和参数
- 优化load_client.rs中的输出格式和数值检查

+++ src/cache.rs
@@ -5,7 +5,6 @@
 //! - 基于容量的LRU淘汰
 //! - 异步刷新
 
-use std::hash::Hash;
 use std::sync::Arc;
 use moka::future::Cache;
 use alloy::primitives::FixedBytes;
@@ -58,7 +57,7 @@ impl CertCache {
         CacheStats {
             hit_count: 0, // moka cache doesn't expose hit/miss counts directly
             miss_count: 0,
-            total_count: entry_count as u64,
+            total_count: entry_count,
             hit_rate: 0.0,
         }
     }
